Separate product and cart test cases into two different files.
